==> Rcode/07_G_matrix_projection_Figure5.R <==
# This code projet all G matrix on the selection surface (gamma eigenvectors) and produce figure 5

rm(list=ls())
gc()
#load libraries
library(coda)
library(matrixStats)
library(data.table)
library(lme4)
library(MCMCglmm)
library(ggplot2)

load('Output_files/RData/VCV_A6140.RData')
load('Output_files/RData/VCV_CA50.RData')
load('Output_files/RData/VCV_CA100.RData')

VCV_mat <- list()
VCV_mat[[1]] <- VCV_mat_A6140[[1]]
for(i in 1:3) VCV_mat[[i+1]] <-  VCV_mat_CA50[[i]]
for(i in 1:3) VCV_mat[[i+4]] <- VCV_mat_CA100[[i]]
rm(VCV_mat_A6140);rm(VCV_mat_CA50);rm(VCV_mat_CA100)

load('Output_files/RData/Analysis_Cemee_Pop_WI_with_gamma_with_Gas_only_Lisbon_with_pop_fixed.RData')


proj_M = eigen(gamma)$vectors
load("Output_files/RData/WI_G-matrix-all.RData")
rm(VCV_mat_A6140)
########## Projection of the G-matrices along the y1-6 #########

temp<-cbind(c(1:7),c("A6140","CA150","CA250","CA350","CA1100","CA2100","CA3100"))
df_for_rot_BP <- NULL

for(i in 1:7){
A_temp= VCV_mat[[as.numeric(temp[i,1])]]$G1_mat/2
A_temp <- I(t(proj_M)%*%A_temp%*%proj_M)
df_for_rot_BP=rbind(df_for_rot_BP,data.frame(pop= temp[i,2],mean_VCV=diag(A_temp),EV=1:6))
}

A_temp= WI_VCV_mat$G1_mat/2
A_temp <- t(proj_M)%*%A_temp%*%proj_M
df_for_rot_BP=rbind(df_for_rot_BP,data.frame(pop= "WI",mean_VCV=diag(A_temp),EV=1:6))

names(df_for_rot_BP)[2] = "Variance"
names(df_for_rot_BP)[3] = "Rotated_traits"

# We have 1.000 MCMC posteriors means, we will split them in 1 x 1000
list_pts=NULL

# Compute all the EV
list_proj_M=list()
all_EV_df=NULL

for(i in 1:ncol(post_mod_gamma_Rdm)){
temp_vect <-post_mod_gamma_Rdm[i,]

temp_vect= temp_vect[c(7:21,1:6)]
rdm_gamma=  matrix(c(temp_vect[16]*2,temp_vect[1:5],
temp_vect[1],temp_vect[17]*2,temp_vect[6:9],
temp_vect[c(2,6)],temp_vect[18]*2,temp_vect[10:12],
temp_vect[c(3,7,10)],temp_vect[19]*2,temp_vect[13:14],
temp_vect[c(4,8,11,13)],temp_vect[20]*2,temp_vect[15],
temp_vect[c(5,9,12,14,15)],temp_vect[21]*2),6,6)

list_proj_M[[i]]= eigen(rdm_gamma)$vectors
all_EV_df=rbind(all_EV_df,eigen(rdm_gamma)$values)
}

rdm_spl=NULL
for(i in 1:400) rdm_spl=rbind(rdm_spl ,sample(1:1000,10))

all_pts_df=NULL
	for(j in 1:7){
temp_pop=c("A6140","CA150","CA250","CA350","CA1100","CA2100","CA3100")[j]
all_Variance=NULL

for(k in 1:nrow(rdm_spl)){
	temp_index=sample(1:length(list_proj_M),1)
	temp_proj_M= list_proj_M[[temp_index]]
  temp_Variance=NULL

		for(r in 1:ncol(rdm_spl)){
	temp_G=t(temp_proj_M)%*%I(matrix(VCV_mat[[j]]$VCV_Mat[rdm_spl[k,r],1:36],6,6)/2)%*% temp_proj_M
temp_Variance=rbind(temp_Variance,diag(temp_G))
}
all_pts_df = rbind(all_pts_df,data.frame(pop=temp_pop,Variance=as.numeric(apply(temp_Variance,2,posterior.mode)),Rotated_traits=1:6,niter=k, EV_gamma= all_EV_df[temp_index,]))
}
}

temp_pop="WI"
all_Variance=NULL

for(k in 1:nrow(rdm_spl)){
	temp_index=sample(1:length(list_proj_M),1)
	temp_proj_M= list_proj_M[[temp_index]]
temp_Variance=NULL

		for(r in 1:ncol(rdm_spl)){

			temp_G=t(temp_proj_M)%*%I(matrix(WI_VCV_mat$VCV_Mat[rdm_spl[k,r],1:36],6,6)/2)%*%temp_proj_M
			temp_Variance=rbind(temp_Variance,diag(temp_G))
}
all_pts_df =rbind(all_pts_df,data.frame(pop=temp_pop,Variance=as.numeric(apply(temp_Variance,2,posterior.mode)),Rotated_traits=1:6,niter=k, EV_gamma= all_EV_df[temp_index,]))
}

df_for_rot_BP$EV_gamma=eigen(gamma)$values

#####

list_pts= all_pts_df
df_for_rot_BP$pop=as.character(df_for_rot_BP$pop)
df_for_rot_BP$pop[df_for_rot_BP$pop=='PB']="PB306"
df_for_rot_BP$pop=factor(df_for_rot_BP$pop,levels=c("A6140","CA150","CA250","CA350","CA1100","CA2100","CA3100","N2","PB306","WI"))
df_for_rot_BP$pop2=c(1,2,2,2,3,3,3,4,5,6,7,8)[as.numeric(df_for_rot_BP$pop)]
list_pts=merge(list_pts,unique(df_for_rot_BP[,c("pop","pop2")]))
vcol=rep(c(grey(.5),"cornflowerblue", "firebrick3","darkgreen","magenta","yellow","cyan","black"),1)

####################################

pdf("plots/Figure5.pdf")


vect_col =c("black", "violetred1","darkorange","firebrick","forestgreen")
vect_y_pops=rep(c(0,100,-33),c(1,3,1))

vect_margin=rbind(c(1.5,5,3.5,0),c(1.5,2.5,3.5,2.5),c(1.5,0,3.5,5),
c(4,5,1,0),c(4,2.5,1,2.5),c(4,0,1,5))
adjust_x=c(0,rep(c(-5,0,5),1),0)

vect_main=c(expression(y[1]),expression(y[2]),expression(y[3]),
expression(y[4]),expression(y[5]),expression(y[6]))
par(mfrow=c(2:3))
for(k in 1:6){

par(mar= vect_margin[k,])
temp_pop=subset(list_pts, Rotated_traits==k & !is.na(Variance))
post_modes_temp=tapply(temp_pop$Variance,temp_pop$pop,posterior.mode)

plot(post_modes_temp[c(1,2,4,6,8)]~vect_y_pops,ylim=c(0.002,0.3),type="n",log="y",xlim=c(-35,120),xlab="",ylab="",bty="n",xaxt="n",yaxt="n",main= vect_main[k],cex.main=2)
if(k==1) mtext(side=2, "Genetic variance",padj=-4,adj=-.6)
Ne_est=1000
lines(c(-33,0),(1-2/(Ne_est))^c(0,173)*(post_modes_temp[8]),lwd=2,lty=2,col="lightgreen")
lines(0:100,(1-2/(Ne_est))^(0:100)*median(post_modes_temp[1]),lwd=2,lty=2)

if(k>3) axis(side=1,at=c(0,50,100),cex.axis=1.2)
if(k %in% c(1,4)){
 axis(side=2,at=c(0.005,0.01,0.05,0.1,0.3),las=1,cex.axis=1.2)
}else{
 axis(side=2,at=c(0.005,0.01,0.05,0.1,0.3),labels=rep("",5),las=1,cex.axis=1.2)
}

vect90=t(matrix(unlist(tapply(temp_pop$Variance,temp_pop$pop,function(x){
	HPDinterval(as.mcmc(x)) })),nrow=2))

vect80=t(matrix(unlist(tapply(temp_pop$Variance,temp_pop$pop,function(x){
	HPDinterval(as.mcmc(x),prob=.8) })),nrow=2))	
		
	arrows((vect_y_pops+adjust_x), vect80[c(1,5:8),1], (vect_y_pops+adjust_x), vect80[c(1,5:8),2],	code=0,length=0,angle=90,col= vect_col,lwd=3)
	arrows((vect_y_pops+adjust_x), vect90[c(1,5:8),1], (vect_y_pops+adjust_x), vect90[c(1,5:8),2],	code=3,length=.07,angle=90,col= grey(.6))

points(post_modes_temp[c(1,2,4,6,8)]~I(vect_y_pops+ adjust_x),pch=16,col= vect_col)

if(k==5) mtext(side=1, "Generation of focal evolution",padj=3)
if(k==4) legend(-10,.01,c("Wild isolates","A6140","CA[1-3]100"),lwd=3,col=c("forestgreen","black","orange"))
}
dev.off()
save(list=ls(),file="Output_files/RData/G_matrix_projection_for_plot_noCA50.RData")
